requirements.txt
asn1crypto==0.24.0
astroid==1.6.6
autopep8==1.4.4
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
configparser==3.7.4
contextlib2==0.5.5
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
Flask==1.1.1
Flask-API==1.1
funcsigs==1.0.2
future==0.17.1
futures==3.1.1
gunicorn==19.9.0
hashID==3.1.4
idna==2.8
importlib-metadata==0.18
ipaddress==1.0.17
isort==4.3.21
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonpickle==1.2
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.1
Markdown==2.6.9
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
parso==0.5.0
pathlib2==2.3.4
pbr==3.1.1
pluggy==0.12.0
prettytable==0.7.2
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.1
Pygments==2.2.0
pylint==1.9.4
pyOpenSSL==17.5.0
python-digitalocean==1.13.2
python-dotenv==0.10.3
python-jsonrpc-server==0.2.0
python-language-server==0.27.0
requests==2.22.0
requests-cache==0.5.0
rope==0.14.0
scandir==1.10.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.12.0
slackclient==1.3.2
snowballstemmer==1.9.0
texttable==0.9.1
typing==3.7.4
urllib3==1.25.3
websocket-client==0.54.0
Werkzeug==0.15.4
vultr==1.0.1
wrapt==1.11.2
yapf==0.27.0
zipp==0.5.2
